- Modify private field naming in PrefManager\n-Move WelcomeActivity class into its own package
